+++ suppaftp/src/async_ftp/mod.rs
@@ -41,7 +41,7 @@ use crate::types::Features;
 ///
 /// It takes a [`SocketAddr`] and returns a [`TcpStream`].
 pub type PassiveStreamBuilder =
-    dyn Fn(SocketAddr) -> Pin<Box<dyn Future<Output = FtpResult<TcpStream>> + Send>>;
+    dyn Fn(SocketAddr) -> Pin<Box<dyn Future<Output = FtpResult<TcpStream>> + Send>> + Send;
 
 /// Stream to interface with the FTP server. This interface is only for the command stream.
 pub struct ImplAsyncFtpStream<T>
@@ -257,7 +257,9 @@ where
     /// to create the [`TcpStream`] for the data connection in passive mode.
     pub fn passive_stream_builder<F>(mut self, stream_builder: F) -> Self
     where
-        F: Fn(SocketAddr) -> Pin<Box<dyn Future<Output = FtpResult<TcpStream>> + Send>> + 'static,
+        F: Fn(SocketAddr) -> Pin<Box<dyn Future<Output = FtpResult<TcpStream>> + Send>>
+            + Send
+            + 'static,
     {
         self.passive_stream_builder = Box::new(stream_builder);
         self
@@ -1445,6 +1447,27 @@ mod test {
                         .map_err(FtpError::ConnectionError)
                 })
             });
+    }
+
+    /// Test if the stream is Send
+    fn is_send<T: Send>(_send: T) {}
+
+    #[async_attributes::test]
+    async fn test_ftp_stream_should_be_send() {
+        crate::log_init();
+        let ftp_stream = AsyncFtpStream::connect("test.rebex.net:21")
+            .await
+            .unwrap()
+            .passive_stream_builder(|addr| {
+                Box::pin(async move {
+                    println!("Connecting to {}", addr);
+                    TcpStream::connect(addr)
+                        .await
+                        .map_err(FtpError::ConnectionError)
+                })
+            });
+
+        is_send::<AsyncFtpStream>(ftp_stream);
     }
 
     // -- test utils